ioengines: merge filecreate, filestat, filedelete engines to fileoperations.c
file operation engines have similar structure, implement them
in one file.
